Add attribute AddCertMounts to service
As per review, we prefer not to hardcode a specific service name in the
dataplane-operator logic for the install-certs role.  Rather, we will
trigger this logic with an AddCertMounts attribute on the service
instead.

A service with this attribute will moount all the TLS certs and keys
for all the services in its openstackAnsibleEE pod.